image_type_tezi.bbclass: refactor image size calculating functions
Drop bootfs_get_size since it has a obvious flaw, it's being called in
both do_image_teziimg and do_image_teziimg_distro but the boot files of
these two tasks are different.

Drop rootfs_get_size, it's trying to get the size of IMAGE_ROOTFS
directory, but in some cases, that is not correct, for instance, when
TEZI_ROOT_SUFFIX is 'ota.tar.gz', the rootfs directory should be
OTA_SYSROOT rather than IMAGE_ROOTFS.

Introduce get_uncompressed_size function, it reads the image size from
a image-size file in ${T} directory, and that image size is written into
this file by the image functions when they making the tarballs.

Also split the duplicated bootfs tarball creating code to a common
function create_bootfs, and it must run as a prefuncs before
rootfs_tezi_run_json and rootfs_tezi_run_distro_json, this ensures the
image size has been written to the file when the later functions run.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>
[use _append for tar commands]
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>