u-boot-distro-boot: run setup for torizon
setup function needs to be run if it's defined in uboot environment,
for instance, loadhdp is being called in it for apalis-imx8 board.

Some setup functions also take care of setting bootargs, which is safe
for torizon as well if we run setup function before bootcmd_args, since
the later will override the bootargs.

Signed-off-by: Ming Liu <ming.liu@toradex.com>