image_type_tezi.bbclass: introduce TEZI_VERSION
Currently we use PV of the image recipe as version number for the
Toradex Easy Installer image json. When trying to implement the new
versioning scheme which introduces timestamps in PV uncovered issues
which are quite complex to overcome, see also this mailing list post:
http://lists.openembedded.org/pipermail/openembedded-core/2019-November/289765.html

To be able to implement the new versioning scheme we need to change
the source of the version number used in the Tezi image json. To be
more dynamic this introduces the variable TEZI_VERSION.

So far the class used the "Software Package Version" without the
(mandatory) date identifier. However, we used the full version with
date indentifier in the file name. This changes to use the full
"Software Package Version" in both cases by default.

Also use IMAGE_CMD_xy[vardepexclude] instead of the task name directly.
This makes sure that image.bbclass does remove the variable from the
image store and avoids premature expansion which leads in basehash
value changed issues.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>