toradex-kernel-localversion: add Toradex specific local version
Add Toradex specific local version implementation for the Linux
kernel. This implementation is a simplified version from what is in
meta-freescale. It can be used with linux-yocto based Linux recipes
as well as regular OE kernel recipes and linux-fslc based recipes.
This allows us to reuse the class in all Linux recipes used by Toradex.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>