u-boot-distro-boot: use lower case for ubi partition name
In 2016.11 U-Boot the MTD partition name which is used in the UBI
distro boot script has been fixed to capital "UBI". In newer U-Boot
this can be configured. With the switch to U-Boot 2019.07 we made
use of that feature to align with the BSP and named the MTD partition
lower case "ubi".

This changes the same in the boot script to make sure Linux looks
for UBI volumes in the correct MTD partition.

Related-to: TOR-237, TOR-513
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>