recipes-core: Add udev-toradex-rules recipe
These rules make sure that a symlink is created for all i2c and tty
devices. This symlinks  are named like on the Toradex carrier boards.
This makes it easy to identify which device in software is matching
the hardware pins.

Additional non-standard devices that are not named clearly on the
carrier board are left out to prevent confusion.

Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>