linux-toradex-mainline: imx 32bit: use a common defconfig and enable can_j1939
Use the common defconfig from the kernel sources toradex-imx_v6_v7_defconfig.
This increases the kernel size for the so far smallest kernel by about 200k.
We need a file for each machine as we do not have a common override to use
in the use-mainline-bsp case.

This additionally adds CAN_J1939 as a kernel module.

Related-to: ELB-3023
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>