u-boot-distro-boot: dont inherit nopackages
The reason we inherit nopackages was that there was no files being
packaged at that moment, but this makes it impossible for a bbappend
to add packages, let's drop it.

Now torizoncore decides to add a ostree-uboot-env package in this
recipe.

Related-to: TOR-1335
Signed-off-by: Ming Liu <ming.liu@toradex.com>
(cherry picked from commit 7b7a2f85e2fcea83c15b1819479506d7d18914bc)