boot.cmd: fix booting from different media
The design of the script allows to use different sources for booting
a kernel, dtb and overlays, and using as a rootfs.
Using devtype to determine the media of rootfs breaks this ability.

Use root_devtype setting to fix this issue. Note the script is still
able to use the same devtype for both boot/rootfs media because both
boot_devtype and root_devtype set from devtype by default.

Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>