boot.cmd: add an ability to use static ip for nfs-root
The kernel is able to use predefined IP address of NFS server.
Add special values of root_devtype "nfs-dhcp" and "nfs-static"
and proceed them accordingly to make the script able to boot the system
from a network without dhcp-server.

To boot the system from the particular server, just set the "serverip"
to the proper IP address of an TFTP/NFS server and set "devtype" to
"tftp".

To use the particular NFS server for mouting the root-fs only, set the
"serverip" to the proper IP address of an NFS server and set
"root_devtype" to "nfs-static".

Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>