device-tree-overlay-filter: initial add
This recipe deploys from the available device tree overlays those which
are applicable for the current machine. Additionally it creates the file
overlays.txt containing a list of overlays which should be started by default.
The files to deploy in the final image are: "overlays.txt overlays/*"

The deploy code is a adapted copy of the image_type_tezi function
tezi_deploy_dt_overlays().

Related-to: ELB-3197
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>