toradex-devicetree.bbclass: introduce bbclass file
The major reason for doing this is because we are currently deploying
all dtbo files to ${DEPLOY_DIR_IMAGE}/devicetree, but not machine
specific ones, and they will be unnecessarily bundled into fitimage
when kernel-fitimage.bbclass is being inherited.

To fix that, we introduce a extra task do_collect_overlays to collect
machine specific overlays to ${DT_FILES_PATH} before do_compile, where
dtb compilation is being compiled in. In this way, only machine
specific overlays would be compiled and deployed.

Now we can drop device-tree-overlay-filter recipe with this change.

Related-to: TOR-1700

Signed-off-by: Ming Liu <ming.liu@toradex.com>