udev: mount.sh: create a symlink to the partition with the bootfiles
Add to the mount.sh script code which creates a symlink to partition
with the bootfiles if that can be guessed from the rootfs partition.
Scan the /proc/cmdline for the rootfs partition. If this is the second
partition of a block device and a first partition exists it creates
the symlink /dev/boot-part.

/etc/fstab consumes the symlink to mount that at /boot.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit c8651feae915eeac23510565dde0f64557696de7)