toradex-sanity.bbclass: introduce bbclass file
I have observed a following OE build issue with the latest
dunfell-5.x.y, as follows:

```
ERROR:  OE-core's config sanity checker detected a potential misconfiguration.
Either fix the cause of this error or at your own risk disable the checker (see sanity.conf).
Following is the list of potential problems / advisories:

TMPDIR changed to include path filtering from the pseudo database.
It is recommended to use a clean TMPDIR with the new pseudo path filtering so TMPDIR would need to be removed to continue.
```

this is due to the ABI VERSION changed in OE, usually we should remove
the TMPDIR and rebuild when seeing this error, but it's not enough for
Toradex builds, since we have split DEPLOY_DIR out of TMPDIR, so we
should also check the ABI VERSION change in DEPLOY_DIR, or else it
might not comply the OE layer.

We should warn the end users about this risk, let's add a toradex
sanity check bbclass for that.

We can also add some other basic Toradex specific sanity checks, for
instance, for fitimage, the end users must inherit toradex-fitimage
rather than kernel-fitimage.

Related-to: ELB-3944

Signed-off-by: Ming Liu <ming.liu@toradex.com>