image_type_tezi: handle emmc being mmcblk2
Note that we currently do not want to use the stable symlinks to the
eMMC for modules were older version of the Toradex Easy Installer exist.
Using the stable symlinks would require a Toradex Easy Instaler 5.3.0 or
newer.

For Verdin iMX8MP 5.3.0 is the first version and thus we can directly use
the stable symlink name pointing to the eMMC.

Related-to: TEI-750
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>