u-boot-distro-boot: fix nfs booting
In the nfs-dhcp case we should not pass any NFS related settings from
U-Boot as the Linux kernel will do its very own DHCP request where it
will get all those settings and knows very well how to apply those.

In the nfs-static case we should add a slash in front of the rootpath as
otherwise an additional /tftpboot/ will be pre-fixed which likely is not
what one wants.

https://man7.org/linux/man-pages/man7/bootparam.7.html
'nfsroot=...'
              This sets the NFS root name to the given string.  If this
              string does not begin with '/' or ',' or a digit, then it
              is prefixed by '/tftpboot/'.  This root name is used in
              case of a net boot.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>