u-boot: bbappend: allow to build boot container from u-boot
U-Boot mainline can build the boot container (imx-boot) from within the
U-Boot tree. In OE this can be enabled with the following override:
imx-boot-container. (imx-base.inc then sets: UBOOT_PROVIDES_BOOT_CONTAINER)
However currently this provides the binary as flash.bin, so we add here
to deploy it additionally under the old name imx-boot

Amend the recipe to allow to enable this.

Related-to: ELB-4243
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>