set-hostname: work around race condition
This service sets the hostname on boot to the machine name amended
by the serial number unless /etc/hostname already exists.

systemd does also set a hostname, if none is available it uses
'localhost' and creates an empty /etc/hostname.

This two mechanism have a race condition so that systemd might
run before set-hostname was able to create /etc/hostname. Likely
only seens a recent systemd update.

Fix that by changing the conditions for executing set-hostname to
run when there is no or an empty /etc/hostname file.

Related-to: ELB-4277
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>