toradex-fitimage.bbclass: introduce FIT_SUPPORTED_INITRAMFS_FSTYPES
It was found when a end user wants to build a squashfs type initramfs
into fitimage, it just fails without printing out any error or warning
messages, which is not right.

Introduce a FIT_SUPPORTED_INITRAMFS_FSTYPES variable to avoid
hard-coding the supported initramfs types, and it could be overridden
in config files. Also break the build when none of a supported
initramfs type is found.

A similar change has been merged into OE, reference:
https://lore.kernel.org/all/20220328141551.571676-1-liu.ming50@gmail.com

Related-to: TEI-1005

Signed-off-by: Ming Liu <ming.liu@toradex.com>