linux-toradex-mainline: Fix Jenkins nightly build
Jenkins recipe uprev (`use-head-next`) is done differently compared to
running manually bitbake on the developer machine.

The hashes are updated in a preliminary step, and afterward bitbake is
running with only the SRCREV variables updated, without setting
use-head-next. Because of that currently the build is failing since
the latest sha from master is not present in the linux-5.18.y branch.

To avoid that just set the KBRANCH to "master" unconditionally, once we
branch out master to kirkstone we will set it to the chosen Linux
Kernel branch.

Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>