linux-toradex-mainline: Fix Jenkins nightly build
Jenkins recipe uprev (`use-head-next`) is done differently compared to
running manually bitbake on the developer machine.

The hashes are updated in a preliminary step, and afterward bitbake is
running with only the SRCREV variables updated, without setting
use-head-next. Because of that currently the build is failing since
KERNEL_VERSION_SANITY_SKIP is not set while building and the version
does not match.

To avoid that just set the KERNEL_VERSION_SANITY_SKIP to "1"
unconditionally, once we branch out master to kirkstone we will need
a to review this.

Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>