linux-toradex-mainline: Lower CMA size percentage on 32-bit SOM
Lower CMA size percentage from 50% to 35% on i.MX 6/6ULL/7 based SOM,
with that the following CMA Memory will be reserved depending on
SKU memory.

| SOM Memory | CMA Memory |
|---         |---         |
| >= 1GiB    |  256MiB    |
| 512MiB     |  192MiB    |
| 256MiB     |  96MiB     |

With this change is possible to have a GPU application and video
playback on low memory SKU, without having allocation failures and
the related OOM killer issues.

No changes on SKU with >= 1GiB memory.

Related-to: ELB-4740
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>