Use `use-head-next` only for SRCREV variable
Remove usage of the `use-head-next` override for anything, apart the
SRCREV variable.

Currently this variable is only used on local builds, while uprev on
Jenkins CI/CD is implemented using `scripts/uprev-srcrev` that update
only the SRCREV variable. This way we ensure a consistent experience
despite how the image is build and we avoid potential confusion.

Related-to: ELB-4885
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>