toradex_imx_v8.config: Do not use -m flag in merge_config.sh
The -m flag causes the script to not call make in the end. That means if
we put e.g. CONFIG_MEDIA_SUPPORT to module, this does not disable all
the
configs below the one until make is being issued.

fix the header that suggests to use this flag in order to enable us to
simplify the diff config by disabling/putting to module whole categories
of configs.

Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>