image_type_tezi: allow for a third boot component
The Verdin AM62 needs a third binary blob in the eMMC boot area.
Add the possibility to specify a third one.

To use it define TPL_BINARY with the name of the file containing the
first boot blob to be stored at OFFSET_BOOTROM_PAYLOAD and define
OFFSET_TPL_PAYLOAD with the offset where the first boot blob expects
the next one.

Not defining TPL_BINARY keeps the image.json creation the same as
before.

Related-to: ELB-5012
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>