image_type_tezi: rename vars related to tpl
The variables TPL_BINARY and OFFSET_TPL_PAYLOAD are not choosen wisely
as they are not used to specify a TPL (Tertiary Program Loader).

They were introduced and used for the Verdin AM62 machine.
The tiboot3 binary is not a TPL (Tertiary Program Loader) but
actually contains the binaries first booted by the boot ROM.

Rename TPL_BINARY and OFFSET_TPL_PAYLOAD to FIRMWARE_BINARY and
OFFSET_FW_PAYLOAD respectively.

Related-to: ELB-5244
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>