linux-toradex-mainline: fix hdmi-csi bridge with iMX6
Revert 1f391df446077d62514ee29f75970dff3d55a432

This reverted commit breaks the async subdevice probing that is needed
by TC358743 HDMI to MIPI-CSI2 bridge on Apalis iMX6. See [1] for details.

The V4L2 async framework for the imx6 seem to be fixed by [2]
patch series, that it is not yet merged.
While waiting for this patch series to be merged we revert this single
commit.

[1] https://lore.kernel.org/all/8360125.31r3eYUQgx@steina-w/
[2] https://lore.kernel.org/all/20230525091615.2324824-1-sakari.ailus@linux.intel.com/

Related-to: ELB-5141
Signed-off-by: Aishwarya Kothari <aishwarya.kothari@toradex.com>