image_type_tezi: select TPL binary based on PID4
Move TPL_BINARY from a variable to a variable flag using the PID4 as a flag and
generate the image json file according to that.
This requires changes on the machine conf where TPL_BINARY is used.

This is required to support different SoC type (GP/HS-FS) that requires a
different tiboot3.bin on the Verdin AM62.

Related-to: ELB-5227
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>
(cherry picked from commit 7fd635f58748064dd034c91605cf6ed25a56a900)