toradex-fitimage.bbclass: follow changed function definition in oe
openembedded-core commit e3eb79c0a1 ("kernel-fitimage: adding support
for Initramfs bundle and u-boot script") is a backport to dunfell.
A parameter was added in the middle of the parameter list of the shell
function fitimage_emit_section_config(). Follow that and change the
calls to it accordingly.

Fixes colibri-imx6ull not booting from the fitimage for the lack of
a default configuration setting.

Related-to: TEI-1053
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>