linux-toradex-mainline-git: extend defconfig for verdin am62
Take the downstream TI kernel toradex_ti_arm64_prune.config as
a blueprint, but drop disabling config options potentially used
for i.MX SoCs.

Rename the file from toradex_imx_v8.config to toradex_aarch64.config
and move it to where the defconfig is stored in the tree
to reflect that it is now used for multiple vendor archs.

Drop addional archs we don't use but still had in the config.

Refresh the defconfig file against kernel 6.5-rc7.

Related-to: ELB-5247
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>