toradex-kernel: move deployment of .config into class
After building the kernel we want the used config being deployed.
This is already implemented for the linux-toradex-mainline recipe,
however the downstream vendor based kernels do not deploy the config.

Move the functionality to a class and inherit that in the mainline
kernel recipe to allow reusing the class in the downstream kernel
recipes.

Related-to: ELB-5149
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit 36705df6b7f3df7d913bc5dcc7299d14d9cd5b9f)