toradex-kernel-localversion: follow oe changes
The setting of localversion has been changed in openembedded-core
to accomodate changes in the kernel sources around v6.3.

This changed that the kernel and external kernel modules with our
way of setting localversion didn't use the same module magic.
One used what we want for LOCALVERSION twice while the other uses
it only once.

Compare with OE-Commits:
commit b378eec156 ("kernel: make LOCALVERSION consistent between recipes")
commit 765b13b730 ("kernel: fix localversion in v6.3+")

Related-to: ELB-5330
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>