u-boot-toradex: Backport mainline patches for nand padding
This commit backports two patches submitted to the U-Boot mainline, both of
which relocate the 'nand_padding' function to the Makefile of 'mach-imx.'

The first patch introduces the 'u-boot-nand.imx' target to the Makefile,
producing a binary file with 1024 leading zeros. The second patch adds a 'dd'
command to the Makefile, appending 10k zeros to the end of the binary.

These changes will eliminate the need of the 'nand_padding' function.

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>
(cherry picked from commit 678579d8783e2d24666ca1b6af58544f67579aa0)