u-boot-toradex: Add patch to fix eMMC dectection on Colibri iMX7D
On Colibri iMX7D 1GB v1.1B, a "new" SoC fusing is being used. This
caused an issue where the GPIO(6,11), used to detect if a board uses
NAND or eMMC, only worked while the module was in recovery mode. Once it
booted from the eMMC, the GPIO was set to output, therefore always
returning 0.

The default behavior should be GPIO return 1 for eMMC device and 0 for a
NAND device, which allows U-Boot to behave correctly.

This also addresses an issue where Tezi 6 was failing to boot because of
wrong variant detection.

Related-to: ELB-5496
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>