image_type_tezi: allow populating "filelist" field
The "filelist" property allows extra files to be copied into the target
filesystem at image installation time. With this implementation we let
users fill in that property at build-time; this is done by settting the
value of the OE variable TEZI_ROOT_FILELIST. The files referenced by
that variable will also become part of the Toradex Easy Installer
image.

Relate-to: TOR-2915

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>