udev-toradex-rules: Add systemd link files to include altnames for CAN
Introduce systemd link files to incorporate alternative names for CAN
interfaces, adding compatibility with names provided in datasheets and
device tree comments.

With this modification, the alternative name for can0 will be 'verdin-can1',
and for can1, 'verdin-can2' on i.MX processors. For the AM62, can1 will
become 'verdin-can1', and can0 will become 'verdin-can2'. The alternative
names will be displayed as follows:

```
4: can0: <NOARP,ECHO> mtu 16 qdisc noop state DOWN group default qlen 10
    link/can
    altname verdin-can1
5: can1: <NOARP,ECHO> mtu 16 qdisc noop state DOWN group default qlen 10
    link/can
    altname verdin-can2
```

Related-to: ELB-5600
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>