u-boot-distro-boot: set kernel cmdline using APPEND variable
Do not depend on U-Boot environment for setting the Linux kernel command
line from the boot script.

Use standard APPEND variable to specify the kernel command line, stop
using ${defargs}, ${setup} and ${vidargs} from U-Boot environment.

APPEND is the standard OE variable to set the kernel command line, it is
for example used in OE-core rootfs-postcommands.bbclass to add "ro" to
the command line and part of the official documentation.

U-Boot ${tdxargs} is still supported and its content is added to the
command line.

Remove consoleblank=0 since this is already the kernel default.

Link: https://docs.yoctoproject.org/dev/kernel-dev/faq.html#how-do-i-change-the-linux-kernel-command-line
Related-to: ELB-5619
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>