set-hostname: update sethostname script to handle read-only root filesystem
The sethostname script has issues when the root filesystem is mounted
as read-only, making not possible to create or modify the /etc/hostname
file.

Therefore, this change updates the sethostname script and adds the
@@HOSTNAMECTL_OPT@@ that can further be replaced with the "--transient"
flag for hostnamectl if the system uses a read-only root FS. This flag
will change the hostname without creating the /etc/hostname file.

With this change, update recipe version from 1.0 to 1.1.

Related-to: ELB-5713
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>