linux-toradex-upstream_6.6: fix filesextrapaths
The immediate evaluation of ${PV} fails here and leads to a not
existing extra search path entry.
Thus linux-toradex-upstream-6.6 is searched after
linux-toradex-upstream and thus files which can be found under both
directories are taken from linux-toradex-upstream.
Fix that by not relying on ${PN} or ${PV} and giving the verbatim
directory name.
That also allows to remove setting FILESEXTRAPATHS in the linux rt
recipe.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>