u-boot: distro-boot-script: allow using complete console vars
The console variable is supposed to be a valid and complete Linux kernel
console parameter, however for legacy reason on some platform (i.MX and
Toradex one) the baudrate is not present.

Currently the boot script always adds the baudrate variable, but this
will break if we fix the console variable to include it.

Fix it adding the baudrate only when there is no "," into the console
variable, this way the bootscript will be able to handle both the
situations.

Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>