linux-firmware: cope with moved directory for unpack
OE master (styhead) no longer unpacks in ${WORKDIR} but rather does
it in path defined by the new variable UNPACKDIR.
A directory created under UNPACKDIR which matches the setting of ${S}
is then moved to ${S}.

If ${S} already exist the move will create an additional subdirectory
rather than creating ${S} with the expected content.

Change the recipe so that ${S} does not already exist from the git
checkout of imx-firmware and thus have the files from linux-firmware*tar
where they are expected.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>