weston-init: replace sleep with udevadm wait
We see on Verdin AM62 and Verdin iMX8MM that often weston fails as
the drm device is not yet ready. Work around this by waiting for
/dev/dri/card0 before trying to start weston.
Depending on the cpu speed and attached devices, e.g. filesystems
needing fsck a fixed but short wait did not catch all cases.

Related-to: ELB-5782
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>