weston-init: wait for card1 if card0 is a gpu
We see on Verdin AM62 and Verdin iMX8MM that often weston fails as
the drm device is not yet ready. Work around this by waiting for
/dev/dri/card0 before trying to start weston.
Then check if card0 is GPU which on its own is not used by weston
and if so wait additionally for card1.

Related-to: ELB-5782
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit 8250eed0c5c981e94cb950b0ee35b107128bda19)