tezi-metadata: wrapup.sh: remove onetime configuration code
Use the generic wrapup.sh script which does nothing also for Apalis/Colibri iMX6.
For Apalis/Colibri iMX6 the wrapup script did fuse/configure the SoC and eMMC
for fastboot mode. Since 2017 Q2 this is the factory default so the
configuration is already done. See the module revision history for exact version
which included the change.

Related-to: ELB-5901
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>