u-boot-distro-boot: accept an initial set of FIT configurations
With this addition, a new OE variable named FITCONF_FDT_OVERLAYS allows
the initial set of FIT configurations to be specified; those
configurations are passed to the bootm command when booting a kernel
from a FIT image. By setting the said variable one could, for example,
ensure some overlays are always applied independent of them being
specified in the overlays.txt file.

Related-to: TOR-3558

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>