boot.cmd: break down rootfs args into static and dynamic parts
Break down the rootfs arguments into a static (known at build-time) and
a dynamic part (defined at runtime) and move the dynamic part to the end
of the bootargs string passed to the kernel. The purpose of this is
making the kernel arguments compatible with the "kernel command-line
protection" hardening feature which will check the initial part of the
kernel command-line against information stored in a signed kernel FIT
image and will validate the dynamic part before starting the kernel.

Related-to: TOR-3558

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>