u-boot-distro-boot: append baudrate to console bootarg conditionally
Introduce the OE variable DISTRO_BOOT_BAUDRATE_MISSING which can be set
to different values to define the behavior of the boot script regarding
how it sets the "console" kernel command-line argument. When set to
"1"/"0" the boot script will assume the baudrate is missing/present in
the "console" variable (i.e. the decision is at build-time) and when set
to "AUTO" the boot script will try to determine the presence of the
baudrate at runtime.

Related-to: TOR-3558

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>